Virtual FE slow controls channel operation, for DTC and ROC, is working

+++ otsdaq-mu2e-calorimeter/FEInterfaces/ROCCalorimeterInterface.h
@@ -24,7 +24,7 @@ class ROCCalorimeterInterface : public ROCPolarFireCoreInterface
 
 	// write and read to registers
 	virtual void 							writeEmulatorRegister	(uint16_t address, uint16_t data_to_write) override;
-	virtual int 							readEmulatorRegister	(uint16_t address) override;
+	virtual uint16_t						readEmulatorRegister	(uint16_t address) override;
